SFR-539 Add new fields to source block
To help researchers identify the source for copyright entries the following fields (already stored in the database) are added to the API response:

- volume number
- group
- matter
- number (for 3rd series entries only)

The API checks to see if `3` appears in the `series` field and if so adds the `number` field.

This also adds these fields to the object definition in the Swagger documentation.