SFR-537 Add LCCN endpoint to CCE API
This adds a LCCN search endpoint to the API for this project and includes any LCCNs that are attached to CCE records in the API responses. These values are queried in the ElasticSearch index and fetched from the database along with the rest of the data when the response fields are gathered.